refactor : 구조 변경 및 주석 제거
